refactor: move rest of manager to create-broadcast-channel.ts
